pyproject.toml
[tool.poetry]
name = "seoul-walk-platform"
version = "0.1.0"
description = "서울시 산책 경로 추천 서비스"
authors = ["Seoul Walk Team"]
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.35.0"
pydeck = "^0.9.0"
folium = "^0.17.0"
streamlit-folium = "^0.21.0"
pandas = "^2.2.0"
geopandas = "^0.14.0"
shapely = "^2.0.0"
requests = "^2.32.0"
httpx = "^0.27.0"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
openai = "^1.30.0"
python-dotenv = "^1.0.0"
osmnx = "^1.9.1"      # OSM 데이터 파이프라인 필수 추가!

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
